Update Artifacts to use lookup tables and reference data from (#495)
`rrtmgp-data-v1.8.1`

Note: The reference results for the longwave solver are generated using a non-scattering solver with rescaling, which differ from the results generated by the TwoStream currently used. Due to the above difference, the longwave tests currently have higher error and are changed to "broken" state. These will be fixed once the Non-scattering solver with rescaling is added.
`PADE` cloud simulations are also commented out and will be added back once appropriate reference data is available for comparison.