Improve readability of map!(x -> x)
A pattern like
```
map!(x -> x, view(flux_up_sw, :, gcol), view(flux_up, :, gcol))
```
is quite hard to understand for what it does. It takes a couple of
minutes to understand that it is just an assignment and it equivalent to
```
flux_up_sw[:, gcol] = flux_up[:, gcol]
```

I removed these instances and replaced them with a clear assignment.

On CPU, this is 1 % slower, but I think that well worth for the added
readability.

```
julia> @benchmark map!(x -> x, view(A, :, 1), view(B, :, 1))
BenchmarkTools.Trial: 10000 samples with 6 evaluations.
 Range (min … max):  4.966 μs …  12.451 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     5.037 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   5.063 μs ± 141.562 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

     ▂▇█▅▁
  ▁▂▅█████▆▄▃▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▂
  4.97 μs         Histogram: frequency by time        5.69 μs <

 Memory estimate: 128 bytes, allocs estimate: 4.

julia> @benchmark view(A, :, 1) .= view(B, :, 1)
BenchmarkTools.Trial: 10000 samples with 6 evaluations.
 Range (min … max):  4.913 μs …  13.299 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     5.086 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   5.121 μs ± 194.785 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

        ▄▇██▇▅▃▁                                              ▂
  ▅▃▃▁▄████████████▇▆▄▃▄▅▆▇▇▆▆▆▆▇▆▇▇▆▅▆▆▆▆▇▇▅▆▅▅▆▅▆▆▅▆▇▄▆▆▅▄▆ █
  4.91 μs      Histogram: log(frequency) by time      6.04 μs <

 Memory estimate: 224 bytes, allocs estimate: 6.
```